requirements.txt
asgiref==3.8.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.4
dj-rest-auth==7.0.2
Django==5.1.6
django-allauth==65.14.1
django-ckeditor==6.7.3
django-cors-headers==4.9.0
django-dotenv==1.4.2
django-filter==24.3
django-js-asset==3.1.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==25.0.3
idna==3.11
oauthlib==3.3.1
packaging==26.0
pycparser==3.0
PyJWT==2.11.0
requests==2.32.5
sqlparse==0.5.5
urllib3==2.6.3
whitenoise==6.11.0
